refact: 관심 질문, 관심 리뷰 엔티티 추가 및 다양한 쿼리 추가

+++ be/zzuag/zzuag-external-api/src/main/java/com/recycle/api/question/dto/response/QuestionByUserResponse.java
@@ -7,9 +7,9 @@ import lombok.Builder;
 public record QuestionByUserResponse(
         Long questionId,
         String content,
-        Long questionLikeCnt,
-        Long totalReviewLikes,
-        Long reviewCount
+        int questionLikeCnt,
+        int totalReviewLikes,
+        int reviewCount
 ) {
     public static QuestionByUserResponse convert(QuestionWithReviewLikesByUserDTO dto) {
         return QuestionByUserResponse.builder()
